// cosmowcl/src/app/shared/models/api/cosmobusiness/operazioneAsincrona.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * cosmobusiness
 * Api per cosmobusiness
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MessaggioOperazioneAsincrona } from './messaggioOperazioneAsincrona.d';


export interface OperazioneAsincrona { 
    uuid: string;
    nome?: string;
    stato: string;
    versione: number;
    risultato?: object;
    dataAvvio?: string;
    dataFine?: string;
    errore?: string;
    dettagliErrore?: string;
    messaggi?: Array<MessaggioOperazioneAsincrona>;
    steps?: Array<OperazioneAsincrona>;
}